chore: move deployment config to project setting page
